feat: Implement Meal Plan and Notification features
- Added MealPlanItemUpsertRequest, MealPlanStatusRequest, MealPlanUpsertRequest, NotificationSettingUpsertRequest, ScheduleMealReminderRequest, and SchedulePrepReminderRequest DTOs.
- Created MealPlanService and NotificationService to handle business logic for meal plans and notifications.
- Introduced IMealPlanService and INotificationService interfaces for service abstraction.
- Updated RecipeController and RecommendationController with additional comments for clarity.
- Enhanced Notification and NotificationSetting entities with new properties and configurations.
- Implemented UnitOfWork pattern to manage MealPlan and Notification repositories.
- Added new endpoints for meal plan and notification management in the API.
- Updated existing services to accommodate new meal plan and notification functionalities.

+++ backend/MenuGreen.BusinessLogicLayer/Services/NutritionTrackingService.cs
@@ -202,7 +202,7 @@ namespace MenuGreen.BusinessLogicLayer.Services
                 ProteinDeviation = totalProtein - targetProtein,
                 CarbsDeviation = totalCarbs - targetCarbs,
                 FatDeviation = totalFat - targetFat,
-                HasWarning = Math.Abs((double)(totalCalories - targetCalories)) > Math.Max(100, targetCalories * 0.10m),
+                HasWarning = Math.Abs((double)(totalCalories - targetCalories)) > (double)Math.Max(100, targetCalories * 0.10m),
                 MealLogs = logs.OrderBy(x => x.LoggedAt).Select(Map).ToList()
             };
         }
